Refactor pools handler for real tracking tests
