+++ app/api/pools/route.test.js
@@ -1,86 +1,90 @@
 // Test for pools API route - showcase leader tracking bug
-// This test documents the bug where userId is not passed to trackBulkGenerations
-//
 // BUG: When creating a sealed pool, showcase leaders are tracked in card_generations
 // but the user_id is not passed, so it's stored as NULL. This causes the showcase
 // leaders to not appear in the user's showcase collection.
 
-import { describe, it } from 'node:test'
+import { describe, it, mock } from 'node:test'
 import assert from 'node:assert'
+import { createPostHandler } from './route-handler.js'
 
-describe('POST /api/pools - showcase leader tracking bug', () => {
-  it('verifies that trackBulkGenerations is called WITH userId after fix', () => {
-    // This test verifies the FIX for the bug in app/api/pools/route.js
-    //
-    // The fix adds userId to the tracking options at lines 169-189
-    // and lines 190-204 (both code paths)
+const baseCards = [
+  { id: 'card-1', name: 'Test Card 1' },
+  { id: 'card-2', name: 'Test Card 2' },
+]
 
-    // Simulate the FIXED behavior (with userId)
-    const userId = 'test-user-123'
-    const fixedTrackingOptions = {
-      packType: 'booster',
-      sourceType: 'sealed',
-      sourceId: 123,
-      sourceShareId: 'abc123',
-      packIndex: 0,
-      userId: userId  // THIS IS NOW INCLUDED IN THE FIX
-    }
+function buildHandler({ body, userId = 'test-user-123' }) {
+  const trackBulkGenerations = mock.fn(async () => {})
+  const query = mock.fn(async () => ({
+    rows: [
+      {
+        id: 123,
+        share_id: body.shareId ?? 'generated-share-id',
+        created_at: '2024-01-01T00:00:00Z',
+      },
+    ],
+  }))
 
-    // Verify that userId is now present
-    assert.strictEqual(
-      fixedTrackingOptions.userId,
-      userId,
-      'Fixed implementation should include userId'
-    )
-
-    // Verify that the fix prevents the showcase bug
-    assert.ok(
-      fixedTrackingOptions.userId !== undefined,
-      'userId must be defined to allow querying showcase leaders by user'
-    )
+  const POST = createPostHandler({
+    query,
+    requireAuth: mock.fn(() => ({ id: userId })),
+    generateShareId: mock.fn(() => 'generated-share-id'),
+    parseBody: mock.fn(async () => body),
+    validateRequired: mock.fn(() => {}),
+    getSetConfig: mock.fn(() => ({ setName: 'Test Set' })),
+    trackBulkGenerations,
+    jsonResponse: mock.fn((data, status) => ({ data, status })),
+    handleApiError: mock.fn((error) => {
+      throw error
+    }),
+    logger: { error: mock.fn() },
   })
 
-  it('verifies the fix applies to both packs array and cards array code paths', () => {
-    // The fix must be applied in TWO places in app/api/pools/route.js:
-    //
-    // 1. Lines 169-189: When packs array is provided
-    //    trackingRecords.push({ card, options: { ...userId: userId... } })
-    //
-    // 2. Lines 190-204: When only cards array is provided (fallback)
-    //    trackingRecords = cards.map(card => ({ card, options: { ...userId: userId... } }))
-    //
-    // Both code paths are now fixed!
+  return { POST, trackBulkGenerations }
+}
 
+describe('POST /api/pools - showcase leader tracking bug', () => {
+  it('calls trackBulkGenerations with userId for packs array path', async () => {
     const userId = 'test-user-123'
-
-    // Fixed behavior for packs array path
-    const packsPathOptions = {
-      packType: 'booster',
-      sourceType: 'sealed',
-      sourceId: 123,
-      sourceShareId: 'abc123',
-      packIndex: 0,
-      userId: userId  // NOW INCLUDED
+    const body = {
+      setCode: 'TST',
+      shareId: 'share-123',
+      cards: baseCards,
+      packs: [baseCards],
     }
 
-    // Fixed behavior for cards array path
-    const cardsPathOptions = {
-      packType: 'booster',
-      sourceType: 'sealed',
-      sourceId: 123,
-      sourceShareId: 'abc123',
-      packIndex: null,
-      userId: userId  // NOW INCLUDED
+    const { POST, trackBulkGenerations } = buildHandler({ body, userId })
+
+    await POST({})
+
+    assert.strictEqual(trackBulkGenerations.mock.calls.length, 1)
+    const [records] = trackBulkGenerations.mock.calls[0].arguments
+
+    assert.strictEqual(records.length, baseCards.length)
+    records.forEach(record => {
+      assert.strictEqual(record.options.userId, userId)
+      assert.strictEqual(record.options.packIndex, 0)
+    })
+  })
+
+  it('calls trackBulkGenerations with userId for cards fallback path', async () => {
+    const userId = 'test-user-456'
+    const body = {
+      setCode: 'TST',
+      shareId: 'share-456',
+      cards: baseCards,
     }
 
-    // Both paths now include userId
-    assert.strictEqual(packsPathOptions.userId, userId)
-    assert.strictEqual(cardsPathOptions.userId, userId)
+    const { POST, trackBulkGenerations } = buildHandler({ body, userId })
 
-    // Verify both paths will allow querying by user_id
-    assert.ok(
-      packsPathOptions.userId && cardsPathOptions.userId,
-      'Both packs and cards array paths must include userId'
-    )
+    await POST({})
+
+    assert.strictEqual(trackBulkGenerations.mock.calls.length, 1)
+    const [records] = trackBulkGenerations.mock.calls[0].arguments
+
+    assert.strictEqual(records.length, baseCards.length)
+    records.forEach(record => {
+      assert.strictEqual(record.options.userId, userId)
+      assert.strictEqual(record.options.packIndex, null)
+    })
   })
 })
